Move DOS contest updates into own module.
